rename `boto_tests` to `botocore_tests`
