+++ tests/botocore_tests/helpers.py
