+++ tests/botocore_tests/unit/cfg/foo_config
