+++ tests/botocore_tests/unit/test_eventstream.py
