+++ tests/botocore_tests/unit/test_handlers.py
