+++ tests/botocore_tests/unit/test_httpchecksum.py
