+++ tests/botocore_tests/unit/test_utils.py
