+++ tests/python3.8/botocore_tests/unit/test_tokens.py
